Revert "[Audiocall] feat: add functionality for auth users"
