+++ src/components/audio-call-game/createMenuGame.ts
@@ -1,33 +1,35 @@
-export default function createMenuGame(previousBookStudy: boolean): HTMLElement {
+export default function createMenuGame(): HTMLElement {
   const elem = document.createElement('div');
-  elem.className = 'audio-call-game container';
-  const meaningfulName1 = `
-  <h1 class="page-heading"><span class="page-heading__rslang">RSLang</span> Audio-call</h1>
-  <div class="buttons-level-block">
-    <h2>Levels</h2>
-    <p>From Easy to most Hard</p>
-    <ol class="btn-wrapper level-menu">
-      <li><button class="btn btn-primary btn-select-level" data-level="0">1</button></li>
-      <li><button class="btn btn-primary btn-select-level" data-level="1">2</button></li>
-      <li><button class="btn btn-primary btn-select-level" data-level="2">3</button></li>
-      <li><button class="btn btn-primary btn-select-level" data-level="3">4</button></li>
-      <li><button class="btn btn-primary btn-select-level" data-level="4">5</button></li>
-      <li><button class="btn btn-primary btn-select-level" data-level="5">6</button></li>
-      <li><button class="btn btn-danger btn-select-level" data-level="6">7</button></li>
-    </ol>
-    <p>
-      Game controls: you can play with both, mouse/touchpad and keyboard.<br>
-      Use keys from 1 to 5 to answer questions and Enter key to activate 'I do not know' button and skip to the next question.<br>
-      To hear the given word press Spacebar.
-    <p>
-  </div>
+  elem.className = 'audio-call-game';
+
+  elem.innerHTML = `
+    <div class="wrapper">
+      <h2>Audio-call mini-game</h2>
+      <div class="btn-group level-menu" role="group" aria-label="Basic radio toggle button group">
+        <input data-level="0" type="radio" class="btn-check" name="btnradio" id="btnradio0" autocomplete="off" checked>
+        <label class="btn btn-outline-primary" for="btnradio0">0</label>
+
+        <input data-level="1" type="radio" class="btn-check" name="btnradio" id="btnradio1" autocomplete="off">
+        <label class="btn btn-outline-primary" for="btnradio1">1</label>
+
+        <input data-level="2" type="radio" class="btn-check" name="btnradio" id="btnradio2" autocomplete="off">
+        <label class="btn btn-outline-primary" for="btnradio2">2</label>
+
+        <input data-level="3" type="radio" class="btn-check" name="btnradio" id="btnradio3" autocomplete="off">
+        <label class="btn btn-outline-primary" for="btnradio3">3</label>
+
+        <input data-level="4" type="radio" class="btn-check" name="btnradio" id="btnradio4" autocomplete="off">
+        <label class="btn btn-outline-primary" for="btnradio4">4</label>
+
+        <input data-level="5" type="radio" class="btn-check" name="btnradio" id="btnradio5" autocomplete="off">
+        <label class="btn btn-outline-primary" for="btnradio5">5</label>
+
+        <input data-level="6" type="radio" class="btn-check" name="btnradio" id="btnradio6" autocomplete="off">
+        <label class="btn btn-outline-primary" for="btnradio6">HARD</label>
+      </div>
+    </div>
     <div class="game-window game-window--hidden"></div>
-  </div>`;
-  const meaningfulName2 = `
-    <h1 class="page-heading"><span class="page-heading__rslang">RSLang</span> Audio-call</h1>
-    <div class="game-window"></div>
+  </div>
   `;
-  elem.innerHTML = previousBookStudy ? meaningfulName2 : meaningfulName1;
-  console.log(elem);
   return elem;
 }
