--- src/components/audio-call-game/game/startNewGame/addEventsForChoiceButtons/hideAnswerButtons/hideAnswerButtons.ts
@@ -1,6 +1,0 @@
-export default function hideAnswerButtons() {
-  const buttonsChoice = document.querySelectorAll('.btn-choice-of-answer');
-  buttonsChoice.forEach((button) => {
-    button.setAttribute('disabled', 'true');
-  });
-}
