--- src/components/audio-call-game/game/startNewGame/addEventsForChoiceButtons/index.ts
@@ -1,4 +1,0 @@
-import hiddenAllButtons from './hideAnswerButtons/hideAnswerButtons';
-import showCurrentWordInfo from './showCurrentWordInfo/showCurrentWordInfo';
-
-export { hiddenAllButtons, showCurrentWordInfo };
