+++ src/components/audio-call-game/menuDifficultyLevel/controlGameWindow/controlGameWindow.ts
@@ -1,6 +1,6 @@
 export default function controlGameWindow(): void {
   const gameWindow = document.querySelector('.game-window');
-  const levelMenu = document.querySelector('.buttons-level-block');
+  const levelMenu = document.querySelector('.level-menu');
   gameWindow?.classList.remove('game-window--hidden');
-  levelMenu?.classList.add('buttons-level-block--hidden');
+  levelMenu?.classList.add('level-menu--hidden');
 }
