+++ src/components/audio-call-game/menuDifficultyLevel/getChoiceOfDifficultyLevel/startNewGame/addEventsForChoiceButtons/showCurrentWordInfo/showCurrentWordInfo.ts
