+++ src/components/audio-call-game/menuDifficultyLevel/getChoiceOfDifficultyLevel/startNewGame/game.types.ts
@@ -4,7 +4,4 @@ export type GameState = {
   correctAnswers: Word[];
   wrongAnswers: Word[];
   currentLevel: number;
-  counterStreakForGame: number;
-  longestStreakForGame: number;
-  newWords: number;
 }
