+++ src/components/audio-call-game/menuDifficultyLevel/getChoiceOfDifficultyLevel/startNewGame/generateWindowGame/getRandomWords/getRandomWords.ts
@@ -1,13 +1,12 @@
 import { Word } from 'types/index';
 
+const MAX_AMOUNT_ANSWERS = 5;
+const MAX_AMOUNT_WORDS_IN_PAGE = 20;
+
 export default function getRandomWords(currentWord: Word, listWords: Word[]): [string, string][] {
-  let maxAmountAnswers = 5;
   const list: [string, string][] = [[currentWord.wordTranslate, currentWord.id]];
-  if (maxAmountAnswers > listWords.length) {
-    maxAmountAnswers = listWords.length;
-  }
-  while (list.length < maxAmountAnswers) {
-    const randomIndex = Math.trunc(Math.random() * listWords.length);
+  while (list.length < MAX_AMOUNT_ANSWERS) {
+    const randomIndex = Math.trunc(Math.random() * MAX_AMOUNT_WORDS_IN_PAGE);
     const newWord = listWords[randomIndex].wordTranslate;
     const idWord = listWords[randomIndex].id;
     if (list.every((el) => el[0] !== newWord)) {
