+++ src/components/audio-call-game/menuDifficultyLevel/getChoiceOfDifficultyLevel/startNewGame/showButtonNextQuestion/showButtonNextQuestion.ts
