+++ src/components/audio-call-page/audio-call-page.css
@@ -1,15 +1,18 @@
-.row-buttons-choice-wrapper:focus {
-  border: none;
-  outline: none;
+.audio-call-page {
+  background-color: var(--bs-cyan);
 }
 
 .current-word {
-  height: 440px;
+  height: 400px;
 }
 
 .button-text-round {
   display: block;
   margin: 0 auto;
+}
+
+.level-menu--hidden {
+  display: none;
 }
 
 .current-word-info--hidden {
@@ -21,95 +24,4 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-}
-
-.level-menu {
-  display: flex;
-  gap: 0.5rem;
-  list-style-type: none;
-  padding: 0px;
-}
-
-.button-wrapper-audiocall {
-  margin-top: 1rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.buttons-level-block {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-}
-
-.page-heading {
-  align-self: flex-start;
-}
-
-.buttons-level-block--hidden {
-  display: none;
-}
-
-.btn-play-again {
-  margin-top: 20px;
-}
-
-.audiocall-speaker-image {
-  width: 3rem;
-  height: 3rem;
-  border-radius: 50%;
-}
-
-.button-audio {
-  width: 5rem;
-  height: 5rem;
-  border-radius: 50%;
-}
-
-.row-buttons-choice-wrapper {
-  margin-top: 24px;
-  height: 60px;
-}
-
-.current-word-info {
-  margin-top: 24px;
-}
-
-.button-audio-result {
-  width: 30px;
-  margin-right: 1.3rem;
-  border: none;
-  background-color: transparent;
-}
-
-.button-audio-image {
-  width: 30px;
-  height: 30px;
-}
-
-.btn-choice-of-answer {
-  font-size: 20px;
-  text-align: center;
-  vertical-align: middle;
-}
-
-.btn-choice-of-answer:disabled {
-  opacity: 1;
-}
-
-.btn-next-question {
-  margin-top: 40px;
-  font-size: 30px;
-}
-
-.wrong-icon-audiocall {
-  color: transparent;  
-  text-shadow: 0 0 0 white;
-}
-
-.correct-icon-audiocall {
-  color: transparent;  
-  text-shadow: 0 0 0 white;
 }
