+++ src/components/audio-call-page/audio-call-page.ts
@@ -1,5 +1,5 @@
 import './audio-call-page.css';
-import { getNewWindowGame } from '../audio-call-game/game/startNewGame/game';
+import { getNewWindowGame } from '../audio-call-game/menuDifficultyLevel/getChoiceOfDifficultyLevel/startNewGame/game';
 
 export default async function getAudioCallPage(): Promise<HTMLElement> {
   const elem = getNewWindowGame();
